Change docker authentication from HTTPS to SSH
Bi-directional TLS was a mistake. I end up needing to maintain 2
CAs (SSH and TLS) without any real benefit, since docker supports using
SSH in contexts.

I can get rid of the wrapper script that was only there to suppress
spurious warnings when using contexts (since that's very unlikely to be
fixed) and a bunch of ansible which was used to set up the TLS. The
cherry on top is hopefully never having to use openssl's horrible
command line again!