Pass data to docker-compose from terraform config
Works by generating the .env file based on Terraform state. This env
file contains all the configuration needed by docker-compose from
terraform, even things that don't need to be environment variables. The
benefit of this is that if the environment variables are specified
properly in docker-compose.yml (using ${VARIABLE:?}), then it will
refuse to work until the variable is found (i.e. remind me to run
terraform before deploying apps).