Migrate from certbot to caddy for TLS
The main reason for this is that I don't want to leave a DNS API token
on the server. caddy can automatically determine which certificates need
to be fetched and use the HTTP challenge for them.

Migrating murmur was a bit annoying, since it needs TLS too. Instead of
giving mumble access to the caddy certs volume, the cleaner way was to
set up caddy to TLS proxy to murmur.

commit 9d1e4ba1f6dff47642e200d67305cce5ea09d8de
Author: Reeto Chatterjee <reetobratachatterjee@gmail.com>
Date:   Thu Feb 25 03:29:55 2021 +0000

    Stop provisioning cloudflare API key for certbot

commit 946e5e19a151453f325a757a8d792427cec27489
Author: Reeto Chatterjee <reetobratachatterjee@gmail.com>
Date:   Thu Feb 25 03:27:47 2021 +0000

    Remove global server configuration for certbot

commit d30529e4fa4d722d7756b9cb767b218eb08b9ddc
Author: Reeto Chatterjee <reetobratachatterjee@gmail.com>
Date:   Thu Feb 25 03:20:20 2021 +0000

    Migrate to caddy for TLS

    * Provision certificates with caddy
    * Set up Caddy as a TLS proxy for Murmur
    * Use a self-signed certificate for Murmur