Remove murmur
Discord is marginally nicer. This removes the need for caddy-l4.

Mumble with TLS proxy was fundamentally broken. Registered users need
client authentication and TLS proxies can't MITM that (by design), so no
clients could be registered using certificates.

The alternatives were to:
* Use a password for server roles.
* Authenticate client certificates at the proxy and use a hardcoded one
  to the server.

Neither of these are particularly nice options; replacing Mumble with
Discord is a lot easier even if it isn't self-hosted :/